Add libgmp header, plus example
'libgmp.inc' contains most of the integer functions of the multiple precision arithmetic library libgmp created. I skip the rational and float point ones, because there are too many! And anyone can prototype the remaining as needed, and add to the file.
I included an example in the examples folder, showcasing how powerfull this library is. Please check it.